🚚 (donate) rename stripe -> checkout
